node: change auth routing
