--- node/src/routes/api/user.js
@@ -1,5 +1,0 @@
-const router = require('express').Router();
-const mongoose = require('mongoose');
-const User = mongoose.model('User');
-
-module.exports = router;
