+++ node/src/routes/index.js
@@ -1,6 +1,5 @@
 const router = require('express').Router();
 
 router.use('/api', require('./api'));
-router.use('/auth', require('./auth'));
 
 module.exports = router;
